chore(engine): remove engine-config & engine-config-builder

+++ crates/engine/src/engine/execute.rs
@@ -197,7 +197,7 @@ impl<R: Runtime> Engine<R> {
                 }
 
                 if let Some(limit) = self.schema.settings.batching.limit {
-                    if requests.len() > limit {
+                    if requests.len() > (limit as usize) {
                         return self.bad_request_but_well_formed_graphql_over_http_request(
                             &request_context,
                             format_args!("batch size exceeds limit of {}", limit),
